sample setup
