+++ com.unity.renderstreaming/Samples~/Example/Scripts/SampleSetup.cs.meta
